requirements.txt
absl-py==1.4.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.3.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.11
Jinja2==3.1.5
keras==3.5.0
libclang==16.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.13.1
packaging==24.2
pillow==11.1.0
protobuf==4.25.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.19.1
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9
six==1.16.0
tensorboard==2.17.1
tensorboard-data-server==0.7.0
tensorflow==2.17.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.3.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.14.1
